Tune this for development/prototyping on Codespaces
The use of Docker for Mac hides/complicates a lot of subtleties.

It's easier/safer to develop/terst this on Codespaces with the right level of permissions than to use Docker on Mac.

The ultimate test is still on Cloud Run V2. See https://cloud.google.com/blog/products/serverless/cloud-run-jobs-and-second-generation-execution-environment-ga